# Refactor/optimize threshold_multiotsu