## Description

<!-- If this is a bug-fix or enhancement, state the issue # it closes -->
<!-- If this is a new feature, reference what paper it implements. -->

This PR basically consists in getting rid of python for loops in the `threshold_multiotsu` function.

This optimization improves dramatically the performance of this function when the number of classes is less then 5:

### Setup

```python
In [1]: from skimage.filters import threshold_multiotsu, threshold_otsu                                 

In [2]: from skimage.data import moon                                                                   

In [3]: img = moon()                              

In [4]: %timeit threshold_otsu(img)                                                                     
1.21 ms ± 444 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)

```

### Before optimization

```python

In [5]: %timeit threshold_multiotsu(img, 2)                                                             
62.4 ms ± 1.21 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [6]: %timeit threshold_multiotsu(img, 3)                                                             
61.8 ms ± 918 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [7]: %timeit threshold_multiotsu(img, 4)                                                             
79.9 ms ± 1.51 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [8]: %timeit threshold_multiotsu(img, 5)                                                             
1.39 s ± 42.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```

### After optimization

```python

In [5]: %timeit threshold_multiotsu(img, 2)                                                             
2.07 ms ± 3.44 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [6]: %timeit threshold_multiotsu(img, 3)                                                             
2.27 ms ± 3.96 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [7]: %timeit threshold_multiotsu(img, 4)                                                             
19.4 ms ± 3.84 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [8]: %timeit threshold_multiotsu(img, 5)                                                             
1.3 s ± 704 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

```

This optimization make the computation time of `threshold_multiotsu` comparable with `threshold_otsu`. This is may be an opportunity to deprecate the actual implementation of `threshold_otsu` and replace it by a call to  `threshold_multiotsu` with `classes=2` to fix the issue #3980.

## For reviewers

<!-- Don't remove the checklist below. -->
- [ ] Check that the PR title is short, concise, and will make sense 1 year
  later.
- [ ] Check that new functions are imported in corresponding `__init__.py`.
- [ ] Check that new features, API changes, and deprecations are mentioned in
      `doc/release/release_dev.rst`.
- [ ] Consider backporting the PR with `@meeseeksdev backport to v0.14.x`