@pep8speaks 2019-09-16
Hello @rfezzani! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`skimage/filters/thresholding.py`](https://github.com/scikit-image/scikit-image/blob/0df2d083cc05be18f26dfad49b4cbcd29d4368b9/skimage/filters/thresholding.py):

> [Line 1150:40](https://github.com/scikit-image/scikit-image/blob/0df2d083cc05be18f26dfad49b4cbcd29d4368b9/skimage/filters/thresholding.py#L1150): [E226](https://duckduckgo.com/?q=pep8%20E226) missing whitespace around arithmetic operator