@alexdesiqueira 2019-09-17
@rfezzani thank you for your help! I will check this PR in a few, ok? It looks promising!

> This is may be an opportunity to deprecate the actual implementation of threshold_otsu and replace it by a call to threshold_multiotsu with classes=2 to fix the issue #3980.

I guess this would not solve the issue... `threshold_multiotsu` is having some issues on negative images, and I'm studying how to solve it. It would be good to maintain the two algorithms; `threshold_otsu` can be used to fine-tune `threshold_multiotsu`, when we have issues like that.