@rfezzani 2019-09-17
Thank you @alexdesiqueira and @sciunto for your quick review. I added a small comment for the step 3 as requested ;-)