@rfezzani 2019-09-17
I based my suggestion of deprecating `threshold_otsu` actual implementation on the comment of @hmaarrfk in #3980:
> I found that there often exists two points that give the same result mathematically. The one that would be chosen would be implementation specific. I had to corner case a few of the points in the test.

This suggests that both results are valid...
But it OK to maintain both versions, this just means that it will be hard to close #3980 :disappointed: