@sciunto 2019-09-17
I propose to continue the discussion on a possible merge of the two functions in #3980 with the additional information that performances are similar and @hmaarrfk comment.